The renderer learns to ORDER primitives across types and to CLIP them; a
ScrollView is built on that, with wheel routing and a fading overlay
indicator. 29 commits, 504 tests, 67 goldens unmoved, warning-free.

M3's scope line in the design spec is nine things and is not one milestone.
This is the first piece, chosen because most of the rest depends on it.

**Spec 7.3 stops being a claim and becomes a property.** Scene held one array
per primitive type and sorted WITHIN each, so encode drew every rect then
every glyph and a rect could never occlude text -- Scene.finalize's own
comment said "nothing in this repo can see a glyph painted through a rect."
The draw list makes draw-call count equal type transitions in z-order, which
is now directly assertable. The honest cost is written down: N interleaved
text rows cost 2N runs, not 2.

**Clipping departs from 7.3's letter and meets its argument.** 7.3 rejected
scissor rects to avoid batch breaks; per-primitive mask data in the instance
buffer causes none either. clip_distance cannot clip a rounded container at
all, and rect_fragment's SDF was already there.

**Three things this milestone got wrong and had to correct in public.**

1. A ruling deleted `flexShrink: 0` from ScrollView's content node on a
   reachability argument. The argument was true about `min-height: 0` and
   irrelevant: flexShrink is load-bearing whenever the automatic minimum sits
   BELOW the flex base size, i.e. any content with text. The probe used
   fixed-height boxes where min-content == max-content and so was structurally
   unable to show the case that mattered -- the same corpus-uniformity hazard
   CLAUDE.md records about the 61 empty-div fixtures. It shipped a horizontal
   ScrollView that did not scroll at all. Restored, and pinned by a test with
   wrappable text.
2. This spec contradicted itself: 1 scoped rounded clip corners OUT while 3.1
   justified choosing the fragment mask BECAUSE it could clip a rounded
   container. The mechanism was chosen on a capability the same document
   excluded. Rounded clipping now ships.
3. The scroll offset was clamped on read and unbounded on write, banking an
   invisible dead band that made scrolling stop intermittently. 20 events of
   -37 into 80pt of travel stored 740.

**Two GPU clip mutants passed all 482 tests** before the final review caught
them: ignoring the mask's origin, and dropping its y factor entirely. Every
clip fixture had origin (0,0) and cut in x only -- so the vertical axis, which
is what a vertical ScrollView clips on, was unasserted.

**The human look found four defects that 501 passing tests could not see**,
two of them in code an approving whole-branch review had already read. One is
pre-existing and now divergence 8, the first divergence here where layout is
right and paint is wrong: a shrink-wrapped Text measures a fractional
max-content, roundLayout's double rounding can store up to a point less, and
Text.paint re-shapes at the rounded width and wraps. Still visible in the demo
by decision -- the honest fixes move a node's stored size.

Style.overflow stays in the declared-but-inert table: ScrollView writes it,
the engine reads it nowhere, and silence there would read as implemented.

+++ Sources/MetalUIPlatform/Platform.swift
@@ -27,8 +27,13 @@ public protocol PlatformWindow: AnyObject {
     var onAppearanceChange: ((Appearance) -> Void)? { get set }
     var onClose: (() -> Void)? { get set }
 
-    /// Begin delivering frame ticks. The callback runs on the main actor.
-    func startDisplayLink(_ tick: @escaping () -> Void)
+    /// Begin delivering frame ticks. The callback runs on the main actor and
+    /// receives the display link's timestamp in seconds.
+    ///
+    /// **The timestamp is the link's, not a wall-clock read.** Every element in
+    /// one frame must see the same instant, and `CACurrentMediaTime()` sampled
+    /// per element would not give them one.
+    func startDisplayLink(_ tick: @escaping (Double) -> Void)
     /// Pausing lets an idle window permit display downclocking (spec 4.4).
     func setDisplayLinkPaused(_ paused: Bool)
 }
